pyproject.toml
[project]
name = "simulacro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]
